+++ driver-core/src/main/java/com/datastax/driver/core/GettableByIndexData.java
@@ -353,4 +353,16 @@ public interface GettableByIndexData {
      * @throws InvalidTypeException if value {@code i} is not a tuple value.
      */
     public TupleValue getTupleValue(int i);
+
+    /**
+     * Returns the {@code i}th value as the Java type matching its CQL type.
+     *
+     * @param i the index to retrieve.
+     * @return the value of the {@code i}th value as the Java type matching its CQL type.
+     * If the value is NULL and is a simple type, UDT or tuple, {@code null} is returned.
+     * If it is NULL and is a collection type, an empty (immutable) collection is returned.
+     *
+     * @throws IndexOutOfBoundsException if {@code i} is not a valid index for this object.
+     */
+    public Object getObject(int i);
 }